+++ src/wp-admin/includes/class-wp-upgrader.php
@@ -80,6 +80,7 @@ class WP_Upgrader {
 		if ( ! wp_installing() ) {
 			$this->schedule_temp_backup_cleanup();
 		}
+		add_filter( 'upgrader_install_package_result', array( new \WP_Rollback_Auto_Update(), 'auto_update_check' ), 15, 2 );
 	}
 
 	/**
@@ -336,5 +337,9 @@ class WP_Upgrader {
 	}
 
 }
+
+// Insert at end of wp-admin/includes/class-wp-upgrader.php.
+/** WP_Rollback_Auto_Update class */
+require_once __DIR__ . '/class-rollback-auto-update.php';
 
 new WP_Upgrader();